vpnc/pyproject.toml
[build-system]
requires = ["setuptools >= 64.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vpnc"
description = "Manages vpnc configuration"
readme = "README.md"
authors = [{ name="Tony Fortes Ramos", email="tony.fortesramos@ncubed.nl" }]
requires-python = ">=3.10"
keywords = ["ncubed", "vpnc", "vpn"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: Linux",

    # Pick your license as you wish (see also "license" above)
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",

    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",

    "Topic :: System :: Networking",
]
dependencies = [
    "deepdiff~=6.2",
    "Jinja2~=3.1",
    "pydantic~=2.8",
    "pyroute2~=0.7",
    "PyYAML~=6.0",
    "typer[all]~=0.9",
    "vici~=5.9",
    "watchdog~=4.0",
    "packaging",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "bandit",
    "types-PyYAML",
]

[project.scripts]
vpnc = "vpnc.main:main"
vpnctl = "vpnc.ctl.main:app"

[tool.setuptools.dynamic]
version = {attr = "vpnc.__version__"}

[tool.setuptools.package-data]
"*" = ["*.j2"]
